+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Changelog for aim
 - Ported CodeCommit config to the model
 - Refactored S3 to use Application StackGroup
 - CPBD artifacts s3 bucket now uses S3 Resource in NetEnv yaml instead
+- Added redirect to Listner rules in the ALB
+- Converted the ALB's listener and listener rules to dicts from lists
 
 0.6.0.dev0 (2019-06-21)
 -----------------------